+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## v1.2.0
+
+Added info HashMap to expose public transport relation metadata
+
+ - Added: info HashMap attribute to PublicTransport struct
+
 ## v1.1.0
 
 Added status struct to know if the parser applied workarounds